[#20241018] Fix for settings modal height
* Enhances UI by introducing tabs and reorganizing settings categories.
* Adjusts modal height for better usability and viewing experience.
